python/requirements.txt
numpy
scipy
librosa
pytest-benchmark
